fix(windows): resolve MSVC build/runtime blockers on native Windows
This branch's own validation checklist flagged "Windows MSVC x86/x64
build and runtime smoke test" as not yet done -- every fix here is a
real bug that confirms exactly that: it had never actually been
compiled/run on native Windows before.

Build errors:
- src/CMakeLists.txt: -Wno-error=tautological-compare is a GCC/Clang-only
  flag, passed unconditionally to MiniAudioImpl.cpp. Guarded with a
  $<$<NOT:$<CXX_COMPILER_ID:MSVC>>:...> generator expression.
- MuPlatform::CreateWindow collided with the Win32 CreateWindow macro
  (<windows.h> #defines it to CreateWindowW/A) -- a hard syntax error on
  MSVC, invisible on macOS/Linux which have no such macro. Renamed to
  CreatePlatformWindow (zero existing callers).
- CpuUsage.cpp called MuGetSystemInfo(&sysInfo) where sysInfo is a raw
  Win32 SYSTEM_INFO, but MuGetSystemInfo actually takes this project's
  own ER_SystemInfo* (a different struct, no processor-count field at
  all) -- looks like a mechanical rename swept up a call site that
  actually wanted the real Win32 GetSystemInfo. Fixed to call ::GetSystemInfo.
- MuLogger.cpp passed a std::filesystem::path directly where spdlog
  wants a std::string -- no implicit conversion exists on any
  standard-conforming compiler. Added .string().
- WinCompat.h's mu_wchar_to_utf8()/mu_narrow_path() were only ever
  defined in the *non-Windows* branch of this compat shim -- genuinely
  missing on native Windows, not a missing-include problem. Added real
  WideCharToMultiByte-backed implementations to the Windows branch.
- w_BuffScriptLoader.cpp called mu_wchar_to_utf8() without including
  Core/Platform/WinCompat.h.
- ListManager.cpp / WZResult.cpp: both headers were updated for
  "Portable types (Story 7.6.6)" (DWORD -> uint32_t) but the .cpp
  implementations weren't -- invisible on macOS/Linux, where
  WinCompat.h's non-Windows branch typedefs DWORD as uint32_t exactly,
  but a real distinct-type overload mismatch on native Windows, where
  DWORD is unsigned long. Fixed all instances to match the headers.
- UIPopup.cpp: brace-init narrowing error converting UIOPTIONS (enum)
  to DWORD in a designated initializer. Added an explicit cast.

Runtime crash (reproduced and stack-walked, not guessed):
- mu::log's internal state (g_loggers, g_sinks, etc.) was a plain
  namespace-scope global. PacketBindings_ClientToServer.h (auto-
  generated) initializes inline globals via mu::platform::GetSymbol()
  at dynamic static-init time; when the .NET client library handle
  wasn't ready yet, GetSymbol()'s null-handle error path called
  mu::log::Get() -- reaching it before this TU's own statics were
  necessarily constructed (C++ gives no ordering guarantee across
  translation units). g_loggers.find() on that not-yet-constructed
  unordered_map was genuine garbage memory, crashing inside <xhash>
  internals on every launch, before any log output. Fixed by making
  the logger state a function-local static (Meyer's singleton) instead
  of a plain global -- C++ guarantees safe construction on first actual
  use regardless of cross-TU static-init order.

Verified end to end: the app now launches and reaches the login scene
on Windows/MSVC (previously it crashed before producing any log output
at all). A separate, pre-existing SDL_gpu pipeline-creation bug
(reproduces identically on both the Vulkan and D3D12 backends) still
blocks actual rendering -- not touched here, out of scope for a build fix.

+++ src/source/UI/Legacy/UIPopup.cpp
@@ -509,7 +509,9 @@ void CUIPopup::Render()
                     .pos = {320 - m_nInputSize / 2 + 5, static_cast<int>(fPosY) - 2},
                     .size = {m_nInputSize, 14},
                     .textLimit = m_nInputTextLength,
-                    .options = m_InputOptions,
+                    // Brace-init narrowing rules treat enum(int)->DWORD(unsigned long) as
+                    // narrowing even though every UIOPTIONS value fits -- explicit cast avoids it.
+                    .options = static_cast<DWORD>(m_InputOptions),
                 });
                 g_pSingleTextInputBox->GiveFocus();
                 g_pSingleTextInputBox->DoAction();
